Check point 2

+++ app/src/main/java/com/vnbamboo/werewolves/Card.java
@@ -1,30 +1,30 @@
 package com.vnbamboo.werewolves;
 
 public class Card {
-    private int id;
-    private int numOrder;
+    private byte id;
+    private byte numOrder;
     private String name;
     private String path;
 
-    public Card( int id, int numOrder, String name, String path ) {
+    public Card( byte id, byte numOrder, String name, String path ) {
         this.id = id;
         this.numOrder = numOrder;
         this.name = name;
         this.path = path;
     }
 
-    public Card( int id, String name, String path ) {
+    public Card( byte id, String name, String path ) {
         this.id = id;
         this.name = name;
         this.path = path;
         this.numOrder = 0;
     }
 
-    public int getId() {
+    public byte getId() {
         return id;
     }
 
-    public void setId( int id ) {
+    public void setId( byte id ) {
         this.id = id;
     }
 
@@ -40,11 +40,11 @@ public class Card {
         return path;
     }
 
-    public int getNumOrder() {
+    public byte getNumOrder() {
         return numOrder;
     }
 
-    public void setNumOrder( int numOrder ) {
+    public void setNumOrder( byte numOrder ) {
         this.numOrder = numOrder;
     }
